add skipping credits by esc
