+++ Assets/Scenes/CreditsScene.unity
@@ -463,6 +463,7 @@ MonoBehaviour:
   mainMenu:
     m_SceneAsset: {fileID: 102900000, guid: 093f513d7b6f9114589d223f7d6cb5b4, type: 3}
     m_SceneName: MainMenu
+  creditSceneManager: {fileID: 778379884}
   creditsText: {fileID: 763265805}
   creditsTextAsset: {fileID: 4900000, guid: dc049b7a7c4dd6246ac004702713ebec, type: 3}
   scrollRect: {fileID: 417505675}
@@ -1051,9 +1052,9 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 417505678}
-        m_TargetAssemblyTypeName: MovingText, Assembly-CSharp
-        m_MethodName: SkipCredits
+      - m_Target: {fileID: 778379884}
+        m_TargetAssemblyTypeName: CreditSceneManager, Assembly-CSharp
+        m_MethodName: EndFinal
         m_Mode: 1
         m_Arguments:
           m_ObjectArgument: {fileID: 0}
@@ -1130,6 +1131,9 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: bed916dcc0ee2dd4294d2d26fc06e9e6, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  mainMenu:
+    m_SceneAsset: {fileID: 102900000, guid: 093f513d7b6f9114589d223f7d6cb5b4, type: 3}
+    m_SceneName: MainMenu
   cutSceneSoundData: {fileID: 11400000, guid: 6da6e50d8a7b009499f9482a9e686dde, type: 2}
   titreSoundData: {fileID: 11400000, guid: 5c2b8d2d7e2340f4e9d8810c2dc24984, type: 2}
   cutSceneRawImage: {fileID: 914994633}
@@ -1715,6 +1719,7 @@ MonoBehaviour:
   mainMenu:
     m_SceneAsset: {fileID: 0}
     m_SceneName: 
+  creditSceneManager: {fileID: 0}
   creditsText: {fileID: 0}
   creditsTextAsset: {fileID: 0}
   scrollRect: {fileID: 0}
