+++ Assets/Scripts/Credits/CreditSceneManager.cs
@@ -1,9 +1,14 @@
 using UnityEngine;
 using UnityEngine.UI;
 using UnityEngine.Video;
+using UnityEngine.SceneManagement;
+using UnityEngine.InputSystem;
 
 public class CreditSceneManager : MonoBehaviour
 {
+    [Header("Scene")]
+    [SerializeField] private SceneField mainMenu;
+
     [SerializeField] private SoundDataSO cutSceneSoundData;
     [SerializeField] private SoundDataSO titreSoundData;
 
@@ -14,6 +19,26 @@ public class CreditSceneManager : MonoBehaviour
 
     private SoundEmitter cutSceneEmitter;
     private SoundEmitter creditsEmitter;
+
+    private InputsTypes _input;
+
+    private void Awake()
+    {
+        _input = new InputsTypes();
+    }
+
+    private void OnEnable()
+    {
+        _input.Enable();
+        _input.UI.Exit.performed += OnExit;
+    }
+
+    private void OnDisable()
+    {
+        _input.UI.Exit.performed -= OnExit;
+        _input.Disable();
+    }
+
     private void Start()
     {
         creditsPopUp.gameObject.SetActive(false);
@@ -29,4 +54,17 @@ public class CreditSceneManager : MonoBehaviour
         cutSceneRawImage.gameObject.SetActive(false);
         creditsPopUp.SetActive(true);
     }
+
+    private void OnExit(InputAction.CallbackContext ctx)
+    {
+        EndFinal();
+    }
+
+    public void EndFinal(){
+
+        SoundManager.Instance.StopAllSounds();
+        EventBus.Raise(EventType.ResetGameState);
+        SceneManager.LoadScene(mainMenu);
+    }
+    
 }
